[gnmi]: Refine AAA test structure
