+++ tests/common/helpers/syslog_helpers.py
@@ -1,8 +1,16 @@
-import os
+from contextlib import contextmanager
+
 import logging
-import pytest
+import os
+import shlex
 import time
-from scapy.all import rdpcap
+import uuid
+
+import pytest
+from scapy.all import Raw, UDP, rdpcap
+
+from tests.common.helpers.assertions import pytest_assert
+from tests.common.utilities import wait_until
 
 DUT_PCAP_FILEPATH = "/tmp/test_syslog_tcpdump.pcap"
 DOCKER_TMP_PATH = "/tmp/"
@@ -27,6 +35,108 @@ def del_syslog_server(dut, syslog_server_ip, module_ignore_errors=False):
         "sudo config syslog del {}".format(syslog_server_ip),
         module_ignore_errors=module_ignore_errors,
     )
+
+
+@contextmanager
+def capture_remote_syslog(dut, destination, vrf=None):
+    capture_file = "/tmp/syslog-{}.pcap".format(uuid.uuid4().hex)
+    capture_pool = None
+    capture_result = None
+    server_added = False
+
+    try:
+        result = add_syslog_server(dut, destination, vrf=vrf)
+        pytest_assert(result["rc"] == 0, "Failed to add syslog server")
+        server_added = True
+
+        remote_action = 'Target="{}" Port="514"'.format(destination)
+
+        def remote_action_configured():
+            command = "grep -F {} /etc/rsyslog.conf".format(
+                shlex.quote(remote_action)
+            )
+            if vrf:
+                command += " | grep -Fq {}".format(
+                    shlex.quote('Device="{}"'.format(vrf))
+                )
+            else:
+                command += " >/dev/null"
+            return dut.shell(
+                command, module_ignore_errors=True
+            )["rc"] == 0
+
+        pytest_assert(
+            wait_until(30, 1, 0, remote_action_configured),
+            "Host rsyslog did not configure UDP/514 forwarding",
+        )
+
+        capture_command = (
+            "sudo timeout 30 tcpdump -i any -y LINUX_SLL -nn "
+            "-s0 -U -w {} {}"
+        ).format(
+            shlex.quote(capture_file),
+            shlex.quote(
+                "udp and dst host {} and dst port 514".format(destination)
+            ),
+        )
+        capture_pool, capture_result = dut.shell(
+            capture_command,
+            module_async=True,
+            module_ignore_errors=True,
+        )
+
+        def capture_started():
+            return dut.shell(
+                "sudo test $(stat -c %s {}) -ge 24".format(
+                    shlex.quote(capture_file)
+                ),
+                module_ignore_errors=True,
+            )["rc"] == 0
+
+        pytest_assert(
+            wait_until(10, 1, 0, capture_started),
+            "UDP/514 packet capture did not start",
+        )
+        yield capture_result, capture_file
+    finally:
+        if capture_pool is not None:
+            if capture_result.ready():
+                capture_pool.close()
+            else:
+                capture_pool.terminate()
+            capture_pool.join()
+        dut.shell(
+            "sudo rm -f {}".format(shlex.quote(capture_file)),
+            module_ignore_errors=True,
+        )
+        if os.path.exists(capture_file):
+            os.remove(capture_file)
+        if server_added:
+            del_syslog_server(dut, destination, module_ignore_errors=True)
+
+
+def read_syslog_payloads(dut, capture_result, capture_file):
+    pytest_assert(
+        wait_until(35, 1, 0, capture_result.ready),
+        "UDP/514 packet capture did not finish",
+    )
+    capture_status = capture_result.get()
+    pcap_status = dut.shell(
+        "sudo test -s {}".format(shlex.quote(capture_file)),
+        module_ignore_errors=True,
+    )
+    pytest_assert(
+        pcap_status["rc"] == 0,
+        "UDP/514 capture file is empty or missing: {}".format(
+            capture_status
+        ),
+    )
+    dut.fetch(src=capture_file, dest="/tmp/", flat=True)
+    return [
+        bytes(packet[Raw].load).decode("utf-8", errors="replace")
+        for packet in rdpcap(capture_file)
+        if UDP in packet and Raw in packet
+    ]
 
 
 # Before real test, check default route on DUT:
